+++ src/films/dto/index.ts
@@ -1,3 +1,2 @@
 export * from './addFilm.dto';
-export * from './addUserFilmReview.dto';
 export * from './updateFilm.dto';
