+++ src/films/films.controller.ts
@@ -6,14 +6,13 @@ import {
   Param,
   Patch,
   Post,
-  Put,
   UseGuards,
 } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
-import { Roles, User } from 'src/common/decorators';
+import { Roles } from 'src/common/decorators';
 import { Role } from 'src/common/enums';
 import { AccessTokenGuard, RolesGuard } from 'src/common/guards';
-import { AddFilmDto, AddUserFilmReviewDto, UpdateFilmDto } from './dto';
+import { AddFilmDto, UpdateFilmDto } from './dto';
 import { FilmsService } from './films.service';
 
 @ApiTags('Films Endpoints')
@@ -65,60 +64,5 @@ export class FilmsController {
   @Get('coming-soon')
   async getComingSoon() {
     return this.filmsService.getComingSoon();
-  }
-
-  @Put('users/favorites/:filmId')
-  @UseGuards(AccessTokenGuard)
-  async addFavoriteFilm(
-    @User('sub') userId: string,
-    @Param('filmId') filmId: string,
-  ) {
-    return this.filmsService.addFavoriteFilm(userId, filmId);
-  }
-
-  @Delete('users/favorites/:filmId')
-  @UseGuards(AccessTokenGuard)
-  async removeFavoriteFilm(
-    @User('sub') userId: string,
-    @Param('filmId') filmId: string,
-  ) {
-    return this.filmsService.removeFavoriteFilm(userId, filmId);
-  }
-
-  @Put('users/watchlist/:filmId')
-  @UseGuards(AccessTokenGuard)
-  async addFilmToWatchlist(
-    @User('sub') userId: string,
-    @Param('filmId') filmId: string,
-  ) {
-    return this.filmsService.addWatchlistFilm(userId, filmId);
-  }
-
-  @Delete('users/watchlist/:filmId')
-  @UseGuards(AccessTokenGuard)
-  async removeWatchlistFilm(
-    @User('sub') userId: string,
-    @Param('filmId') filmId: string,
-  ) {
-    return this.filmsService.removeWatchlistFilm(userId, filmId);
-  }
-
-  @Put('users/reviews/:filmId')
-  @UseGuards(AccessTokenGuard)
-  async addFilmReview(
-    @User('sub') userId: string,
-    @Param('filmId') filmId: string,
-    @Body() dto: AddUserFilmReviewDto,
-  ) {
-    return this.filmsService.addFilmReview(userId, filmId, dto);
-  }
-
-  @Delete('users/reviews/:filmId')
-  @UseGuards(AccessTokenGuard)
-  async removeFilmReview(
-    @User('sub') userId: string,
-    @Param('filmId') filmId: string,
-  ) {
-    return this.filmsService.removeFilmReview(userId, filmId);
   }
 }
