+++ src/films/films.module.ts
@@ -5,5 +5,6 @@ import { FilmsService } from './films.service';
 @Module({
   controllers: [FilmsController],
   providers: [FilmsService],
+  exports: [FilmsService],
 })
 export class FilmsModule {}
