+++ src/films/films.service.ts
@@ -4,7 +4,7 @@ import {
   NotFoundException,
 } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
-import { AddFilmDto, AddUserFilmReviewDto, UpdateFilmDto } from './dto';
+import { AddFilmDto, UpdateFilmDto } from './dto';
 
 @Injectable()
 export class FilmsService {
@@ -248,192 +248,5 @@ export class FilmsService {
     }
 
     return list;
-  }
-
-  async addFavoriteFilm(userId: string, filmId: string) {
-    // Check if the film exists
-    const film = await this.findById(filmId);
-    if (!film) {
-      throw new NotFoundException('Film not found');
-    }
-
-    // Check if user already added the film to favorites
-    const existingFavorite = await this.prisma.userFilmFavorite.findFirst({
-      where: {
-        user_id: userId,
-        film_id: filmId,
-      },
-    });
-    if (existingFavorite) {
-      throw new BadRequestException('Film already added to favorites');
-    }
-
-    // Add the film to the user's favorite films
-    await this.prisma.userFilmFavorite.create({
-      data: {
-        user_id: userId,
-        film_id: filmId,
-      },
-    });
-
-    // Calculate the total favorites for the film
-    await this.calculateTotalFavorites(filmId);
-  }
-
-  async removeFavoriteFilm(userId: string, filmId: string) {
-    // Check if the film exists
-    const film = await this.findById(filmId);
-    if (!film) {
-      throw new NotFoundException('Film not found');
-    }
-
-    // Check if the user added the film to favorites
-    const existingFavorite = await this.prisma.userFilmFavorite.findFirst({
-      where: {
-        user_id: userId,
-        film_id: filmId,
-      },
-    });
-    if (!existingFavorite) {
-      throw new BadRequestException('Film not added to favorites');
-    }
-
-    // Remove the film from the user's favorite films
-    await this.prisma.userFilmFavorite.delete({
-      where: {
-        id: existingFavorite.id,
-      },
-    });
-
-    // Calculate the total favorites for the film
-    await this.calculateTotalFavorites(filmId);
-  }
-
-  async addWatchlistFilm(userId: string, filmId: string) {
-    // Check if the film exists
-    const film = await this.findById(filmId);
-    if (!film) {
-      throw new NotFoundException('Film not found');
-    }
-
-    // Check if user already added the film to watchlist
-    const existingWatchlist = await this.prisma.userFilmWatchlist.findFirst({
-      where: {
-        user_id: userId,
-        film_id: filmId,
-      },
-    });
-    if (existingWatchlist) {
-      throw new BadRequestException('Film already added to watchlist');
-    }
-
-    // Add the film to the user's watchlist
-    await this.prisma.userFilmWatchlist.create({
-      data: {
-        user_id: userId,
-        film_id: filmId,
-      },
-    });
-  }
-
-  async removeWatchlistFilm(userId: string, filmId: string) {
-    // Check if the film exists
-    const film = await this.findById(filmId);
-    if (!film) {
-      throw new NotFoundException('Film not found');
-    }
-
-    // Check if the user added the film to watchlist
-    const existingWatchlist = await this.prisma.userFilmWatchlist.findFirst({
-      where: {
-        user_id: userId,
-        film_id: filmId,
-      },
-    });
-    if (!existingWatchlist) {
-      throw new BadRequestException('Film not added to watchlist');
-    }
-
-    // Remove the film from the user's watchlist
-    await this.prisma.userFilmWatchlist.delete({
-      where: {
-        id: existingWatchlist.id,
-      },
-    });
-  }
-
-  async addFilmReview(
-    userId: string,
-    filmId: string,
-    dto: AddUserFilmReviewDto,
-  ) {
-    // Check if the film exists
-    const film = await this.findById(filmId);
-    if (!film) {
-      throw new NotFoundException('Film not found');
-    }
-
-    // Check if the user added review to the film
-    const existingReview = await this.prisma.userFilmReview.findFirst({
-      where: {
-        user_id: userId,
-        film_id: filmId,
-      },
-    });
-    if (existingReview) {
-      throw new BadRequestException('Review already added');
-    }
-
-    // Check if the rating is floating point
-    if (!Number.isInteger(dto.rating)) {
-      throw new BadRequestException('Rating must be an integer');
-    }
-
-    // Check if the rating is between 1 and 5
-    if (dto.rating < 1 || dto.rating > 5) {
-      throw new BadRequestException('Rating must be between 1 and 5');
-    }
-
-    // Add the review to the film
-    await this.prisma.userFilmReview.create({
-      data: {
-        user_id: userId,
-        film_id: filmId,
-        rating: dto.rating,
-        review: dto.review,
-      },
-    });
-
-    // Calculate the total rating for the film
-    await this.calculateTotalRating(filmId);
-  }
-
-  async removeFilmReview(userId: string, filmId: string) {
-    // Check if the film exists
-    const film = await this.findById(filmId);
-    if (!film) {
-      throw new NotFoundException('Film not found');
-    }
-
-    // Check if the user added review to the film
-    const existingReview = await this.prisma.userFilmReview.findFirst({
-      where: {
-        user_id: userId,
-        film_id: filmId,
-      },
-    });
-    if (!existingReview) {
-      throw new BadRequestException('Review not found');
-    }
-
-    // Remove the review from the film
-    await this.prisma.userFilmReview.delete({
-      where: {
-        id: existingReview.id,
-      },
-    });
-
-    // Calculate the total rating for the film
-    await this.calculateTotalRating(filmId);
   }
 }
