+++ src/users/dto/index.ts
@@ -1,1 +1,2 @@
+export * from './addFilmReview.dto';
 export * from './updateUser.dto';
