+++ src/users/users.controller.ts
@@ -5,12 +5,13 @@ import {
   Get,
   Param,
   Patch,
+  Put,
   UseGuards,
 } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
 import { User } from 'src/common/decorators';
 import { AccessTokenGuard } from 'src/common/guards';
-import { UpdateUserDto } from './dto';
+import { AddFilmReviewDto, UpdateUserDto } from './dto';
 import { UsersService } from './users.service';
 
 @ApiTags('Users Endpoints')
@@ -40,5 +41,60 @@ export class UsersController {
     @Param('username') username: string,
   ) {
     return this.usersService.deleteProfile(userId, username);
+  }
+
+  @Put('films/favorites/:filmId')
+  @UseGuards(AccessTokenGuard)
+  async addFavoriteFilm(
+    @User('sub') userId: string,
+    @Param('filmId') filmId: string,
+  ) {
+    return this.usersService.addFavoriteFilm(userId, filmId);
+  }
+
+  @Delete('films/favorites/:filmId')
+  @UseGuards(AccessTokenGuard)
+  async removeFavoriteFilm(
+    @User('sub') userId: string,
+    @Param('filmId') filmId: string,
+  ) {
+    return this.usersService.removeFavoriteFilm(userId, filmId);
+  }
+
+  @Put('films/watchlist/:filmId')
+  @UseGuards(AccessTokenGuard)
+  async addFilmToWatchlist(
+    @User('sub') userId: string,
+    @Param('filmId') filmId: string,
+  ) {
+    return this.usersService.addWatchlistFilm(userId, filmId);
+  }
+
+  @Delete('films/watchlist/:filmId')
+  @UseGuards(AccessTokenGuard)
+  async removeWatchlistFilm(
+    @User('sub') userId: string,
+    @Param('filmId') filmId: string,
+  ) {
+    return this.usersService.removeWatchlistFilm(userId, filmId);
+  }
+
+  @Put('films/reviews/:filmId')
+  @UseGuards(AccessTokenGuard)
+  async addFilmReview(
+    @User('sub') userId: string,
+    @Param('filmId') filmId: string,
+    @Body() dto: AddFilmReviewDto,
+  ) {
+    return this.usersService.addFilmReview(userId, filmId, dto);
+  }
+
+  @Delete('films/reviews/:filmId')
+  @UseGuards(AccessTokenGuard)
+  async removeFilmReview(
+    @User('sub') userId: string,
+    @Param('filmId') filmId: string,
+  ) {
+    return this.usersService.removeFilmReview(userId, filmId);
   }
 }
